+++ src/internal/get_mtu.c
@@ -3,13 +3,14 @@
 #include "stdio.h"
 #include <stdatomic.h>
 #include <net/if.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
 #include <sys/socket.h>
 #include <sys/ioctl.h>
 
-uint32_t get_mtu(const char* restrict const interface, const int sockfd) {
+uint16_t get_mtu(const char* restrict const interface, const int sockfd) {
     struct ifreq ifr;
 
 
@@ -21,5 +22,5 @@ uint32_t get_mtu(const char* restrict const interface, const int sockfd) {
         exit(EXIT_FAILURE);
     }
 
-    return ifr.ifr_mtu;
+    return (uint16_t)ifr.ifr_mtu;
 }
