+++ src/make_request.c
@@ -38,21 +38,21 @@ struct SwiftNetClientPacketData* swiftnet_client_make_request(struct SwiftNetCli
     #ifdef SWIFT_NET_BACKEND_DPDK
     packet_length = packet->total_data_size;
     #else
-    packet_length = packet->packet_append_pointer - packet->packet_data_start;
+    packet_length = (uint32_t)(packet->packet_append_pointer - packet->packet_data_start);
     #endif
 
     swiftnet_send_packet(client->maximum_transmission_unit, client->port_info, packet, packet_length, &client->server_addr, &client->packets_sending, &client->packets_sending_memory_allocator, client->eth_header, client->network_data, request_sent, false, 0);
 
 
     gettimeofday(&tv, NULL);
-    start = (uint32_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    start = (uint32_t)(tv.tv_sec * 1000 + tv.tv_usec / 1000);
 
     goto check_response;
 
 
 check_response:
     gettimeofday(&tv, NULL);
-    end = (uint32_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    end = (uint32_t)(tv.tv_sec * 1000 + tv.tv_usec / 1000);
 
     if (start + timeout_ms < end) {
         delete_request_sent(request_sent);
@@ -86,7 +86,7 @@ struct SwiftNetServerPacketData* swiftnet_server_make_request(struct SwiftNetSer
     #ifdef SWIFT_NET_BACKEND_DPDK
     packet_length = packet->total_data_size;
     #else
-    packet_length = packet->packet_append_pointer - packet->packet_data_start;
+    packet_length = (uint32_t)(packet->packet_append_pointer - packet->packet_data_start);
     #endif
 
     port_info = (struct SwiftNetPortInfo){.destination_port = addr_data.port, .source_port = server->server_port};
@@ -94,14 +94,14 @@ struct SwiftNetServerPacketData* swiftnet_server_make_request(struct SwiftNetSer
     swiftnet_send_packet(addr_data.maximum_transmission_unit, port_info, packet, packet_length, &addr_data.sender_address, &server->packets_sending, &server->packets_sending_memory_allocator, server->eth_header, server->network_data, request_sent, false, 0);
 
     gettimeofday(&tv, NULL);
-    start = (uint32_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    start = (uint32_t)(tv.tv_sec * 1000 + tv.tv_usec / 1000);
 
     goto check_response;
 
 
 check_response:
     gettimeofday(&tv, NULL);
-    end = (uint32_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    end = (uint32_t)(tv.tv_sec * 1000 + tv.tv_usec / 1000);
 
     if (start + timeout_ms < end) {
         delete_request_sent(request_sent);
